+++ CHANGELOG.md
@@ -11,6 +11,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+## [0.4.0] - 2022-06-22
+
+### Added
+
+- **GM-82** - Video Section component.
+- **GM-81** - BannerMedium component.
+
+### Fixed
+
 ## [0.3.0] - 2022-06-15
 
 ### Added